WIP refactor/OpenAI Factories
